+++ src/unit_definitions.rs
@@ -23,6 +23,7 @@ system!(
         CatalyticActivityConcentration,
         Curvature,
         DiffusionCoefficient,
+        Dimensionless,
         DoseEquivalent,
         DynamicViscosity,
         ElectricCharge,